requirements.txt
aiosmtplib==2.0.1
alembic==1.11.1
anyio==3.7.0
bcrypt==4.0.1
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
click==8.1.3
cloudinary==1.33.0
colorama==0.4.6
cryptography==41.0.1
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
exceptiongroup==1.1.1
fastapi==0.95.2
fastapi-mail==1.2.8
greenlet==2.0.2
h11==0.14.0
httptools==0.5.0
idna==3.4
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
passlib==1.7.4
phonenumbers==8.13.13
psycopg2==2.9.6
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.8
pyotp==2.8.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.15
SQLAlchemy-Utils==0.41.1
starlette==0.27.0
typing_extensions==4.6.2
urllib3==1.26.16
uvicorn==0.22.0
watchfiles==0.19.0
websockets==11.0.3
